requirements.txt
pillow==10.4.0
djangorestframework==3.15.2
Django==5.1.1
djoser==2.2.3
